+++ internal/web/embed/public/assets/hooks-C-P23sc-.js
@@ -1,2 +1,0 @@
-import{e as t,f as n,g as i,h as o,i as r}from"./index-DUH46VSk.js";function c(){return t()}function h(e){return n(e,{skip:e===void 0})}function g(e){return i(e)}function m(){return o()}function S(){const[e,u]=r();return{...u,isPending:u.isLoading,mutateAsync:s=>e(s).unwrap()}}export{h as a,g as b,m as c,S as d,c as u};
-//# sourceMappingURL=hooks-C-P23sc-.js.map